crap... utils.py
