+++ korgbot.py
@@ -69,37 +69,6 @@ async def on_member_join(member):
         f'If I learn a new trick I\'ll brag about it in my instructions channel.\n'
     )
 
-#Make logging available
-async def log(client, string, timestamp=True):
-    # Log to stdout
-    timestamp_string = ''
-    if timestamp:
-        timestamp_string = f'[{str(datetime.now())[:-7]}]'
-    print(timestamp_string + ' ' + string)
-
-    # Log to channel
-    for guild in client.guilds:
-        for channel in guild.text_channels:
-            if channel.name == 'bot-logs':
-                await channel.send(string)
-
-    # Log to file
-    try:
-        async with aiofiles.open('log', mode='r') as f:
-            previous_logs = await f.readlines()
-    except FileNotFoundError:
-        previous_logs = []
-
-    async with aiofiles.open('log', mode='w') as f:
-        for line in previous_logs:
-            await f.write(line.strip() + '\n')
-        await f.write(timestamp_string + ' ' + string + '\n')
-
-
-#easier direct messages
-async def dm(member, content):
-    channel = await member.create_dm()
-    await channel.send(content)
 
 #error events
 @client.event
